+++ typescript/solver/solvers/index.ts
@@ -1,2 +1,2 @@
-export * as onChain from "./onChain/index.js";
 export * as eco from "./eco/index.js";
+export * as hyperlane7683 from "./hyperlane7683/index.js";
